experiments/cognitive-bias-antidote/requirements.txt
fastapi
uvicorn
google-generativeai
jinja2
python-multipart
pydantic
